refactor: extract shared player POV hit helper
Move Item.getPlayerPOVHitResult out of BottleItem and PlaceOnWaterBlockItem
into item_utils so both use the same clip path.
